Continued refactoring PointOffset.affmapToPixelsMaybeTransposeInContextOf.
The after-call which calculates bar size is way too messy, this needs to
be cleaned up.

core new tests working, but widget tests failing for bar size
calculations were moved to followup method.. This needs to be
fixed.

Also lots of cleanup and simplification needed in after-call of
PointOffset.affmapToPixelsMaybeTransposeInContextOf.